서버 주소 공통변수로 만듦

+++ server/src/socket-server/socket-server.gateway.ts
@@ -1,8 +1,13 @@
 import { WebSocketGateway, WebSocketServer, SubscribeMessage } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
 import { SocketServerService, UserInfo, InputUser } from './socket-server.service';
+import { serverAddress } from 'common/server-common';
 
-@WebSocketGateway()
+@WebSocketGateway({
+  cors: {
+    origin: serverAddress
+  }
+})
 export class SocketServerGateway {
   constructor(private readonly SocketServerService: SocketServerService) {}
   @WebSocketServer()
@@ -16,13 +21,14 @@ export class SocketServerGateway {
     const data = this.SocketServerService.getOnlineUser();
     client.emit('getOnline', data)
   }
-  @SubscribeMessage('outUser')
-  disconnectUser(client: Socket, payload: string): void {
-    this.SocketServerService.deleteOnlineUser(payload);
-  }
   @SubscribeMessage('checkNickName')
   possible(client: Socket, payload: string): void {
     const result: boolean = this.SocketServerService.checkDuplicationNickName(payload);
     client.emit('checkNickName', result);
+  }
+  @SubscribeMessage('disconnect')
+  disconnectUser(client: Socket, payload: string): void {
+    if (payload === '')
+    this.SocketServerService.deleteOnlineUser(payload);
   }
 }
